refactor for pure
