+++ patchsdk/patchsdk/src/main/AndroidManifest.xml
@@ -20,14 +20,5 @@
             </intent-filter>
         </receiver>
 
-        <service
-            android:name=".tinker.SampleResultService"
-            android:exported="false"/>
-
-        <service
-            android:name=".tinker.SamplePatchService"
-            android:exported="false"
-            android:process=":patch"/>
-
     </application>
 </manifest>
