+++ patchsdk/sample/src/main/java/com/dx168/patchsdk/sample/MyApplicationLike.java
@@ -1,15 +1,16 @@
 package com.dx168.patchsdk.sample;
 
 import android.app.Application;
+import android.content.Context;
 import android.content.Intent;
-import android.content.res.AssetManager;
-import android.content.res.Resources;
 import android.util.Log;
 
+import com.dx168.patchsdk.ActualPatchManager;
 import com.dx168.patchsdk.Listener;
 import com.dx168.patchsdk.PatchManager;
-import com.dx168.patchsdk.tinker.SampleApplicationLike;
+import com.dx168.patchsdk.sample.tinker.SampleApplicationLike;
 import com.tencent.tinker.anno.DefaultLifeCycle;
+import com.tencent.tinker.lib.tinker.TinkerInstaller;
 import com.tencent.tinker.loader.shareutil.ShareConstants;
 
 /**
@@ -35,7 +36,17 @@ public class MyApplicationLike extends SampleApplicationLike {
         super.onCreate();
         String appId = "20170112162040035-6936";
         String appSecret = "d978d00c0c1344959afa9d0a39d7dab3";
-        PatchManager.getInstance().init(getApplication(), "http://hotfix.dx168.com/hotfix-apis/", appId, appSecret);
+        PatchManager.getInstance().init(getApplication(), "http://hotfix.dx168.com/hotfix-apis/", appId, appSecret, new ActualPatchManager() {
+            @Override
+            public void cleanPatch(Context context) {
+                TinkerInstaller.cleanPatch(context);
+            }
+
+            @Override
+            public void patch(Context context, String patchPath) {
+                TinkerInstaller.onReceiveUpgradePatch(context, patchPath);
+            }
+        });
         PatchManager.getInstance().register(new Listener() {
             @Override
             public void onQuerySuccess(String response) {
