+++ patchsdk/sample/src/main/java/com/dx168/patchsdk/sample/tinker/SampleApplicationLike.java
@@ -1,11 +1,9 @@
-package com.dx168.patchsdk.tinker;
+package com.dx168.patchsdk.sample.tinker;
 
 import android.annotation.TargetApi;
 import android.app.Application;
 import android.content.Context;
 import android.content.Intent;
-import android.content.res.AssetManager;
-import android.content.res.Resources;
 import android.os.Build;
 import android.support.multidex.MultiDex;
