+++ patchsdk/sample/src/main/java/com/dx168/patchsdk/sample/tinker/SamplePatchListener.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.dx168.patchsdk.tinker;
+package com.dx168.patchsdk.sample.tinker;
 
 import android.app.ActivityManager;
 import android.content.Context;
