+++ patchsdk/sample/src/main/java/com/dx168/patchsdk/sample/tinker/SamplePatchReporter.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.dx168.patchsdk.tinker;
+package com.dx168.patchsdk.sample.tinker;
 
 import android.content.Context;
 import android.content.Intent;
