+++ patchsdk/sample/src/main/java/com/dx168/patchsdk/sample/tinker/SamplePatchService.java
@@ -1,4 +1,4 @@
-package com.dx168.patchsdk.tinker;
+package com.dx168.patchsdk.sample.tinker;
 
 import android.content.Context;
 import android.content.Intent;
